pyproject.toml
# pyproject.toml

[tool.setuptools]
package-dir = {"" = "dydx4"}

[tool.poetry]
name = "dydx4"
version = "0.1.4"
description = "dYdX v4 Client"
authors = ["John Huang <contact@dydx.exchange>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Cython = "^0.29.0"
aiohttp = "^3.8.1"
bech32 = "^1.2.0"
bip_utils = "^2.7.0"
brownie = "^0.5.0"
cytoolz = "^0.12.1"
dateparser = "^1.0.0"
ecdsa = "^0.18.0"
eth_keys = "^0.4.0"
eth-account = "^0.9.0"
protobuf = "^4.23"
grpcio-tools = "^1.56"
grpcio = "^1.56"
mpmath = "^1.3.0"
requests = "^2.31.0"
six = "^1.16"
sympy = "^1.12.0"
web3 = "^6.5.0"
websocket_client = "^1.6.1"

[tool.poetry.dev-dependencies]
wheel = "^0.35.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
